Moved helper functions to separate file in utils folder
